requirements.txt
python-telegram-bot==20.7
pymongo
python-dotenv
unidecode
nudenet
opencv-python-headless
Pillow
